ES6ify, and stop self-registering
