+++ .gitignore
@@ -1,4 +1,5 @@
-.DS_Store
+.*
+!.gitignore
 node_modules/
 tmp/
 coverage.html
