+++ package.json
@@ -1,40 +1,44 @@
 {
-	"name": "ractive-adaptors-backbone",
-	"description": "Backbone adaptor for Ractive.js",
-	"version": "0.2.0",
-	"homepage": "http://ractivejs.org",
-	"main": "./ractive-adaptors-backbone.js",
-	"author": {
-		"name": "Rich Harris"
-	},
-	"licenses": [
-		{
-			"type": "MIT"
-		}
-	],
-	"repository": {
-		"type": "git",
-		"url": "https://github.com/ractivejs/ractive-adaptors-backbone.git"
-	},
-	"peerDependencies": {
-		"backbone": ">= 1.0.0",
-		"ractive": ">= 0.3.9"
-	},
-	"scripts": {
-		"test": "jshint ractive-adaptors-backbone.js && mocha",
-		"coverage": "(cp ractive-adaptors-backbone.js ractive-adaptors-backbone.js~; jscoverage ractive-adaptors-backbone.js; mv ractive-adaptors-backbone-cov.js ractive-adaptors-backbone.js; mocha -R html-cov > coverage.html); mv ractive-adaptors-backbone.js~ ractive-adaptors-backbone.js",
-		"prepublish": "uglifyjs -m < ractive-adaptors-backbone.js > ractive-adaptors-backbone.min.js"
-	},
-	"devDependencies": {
-		"backbone": ">= 1.0.0",
-		"chai": "^1.9.2",
-		"jshint": "^2.5.6",
-		"mocha": "^1.21.4",
-		"mocha-clean": "^0.2.1",
-		"mocha-repeat": "^0.1.0",
-		"proxyquire": "^1.0.1",
-		"ractive": ">= 0.3.9",
-		"uglify-js": "^2.4.15",
-		"underscore": "^1.7.0"
-	}
+  "name": "ractive-adaptors-backbone",
+  "description": "Backbone adaptor for Ractive.js",
+  "version": "0.2.0",
+  "homepage": "http://ractivejs.org",
+  "main": "dist/ractive-adaptors-backbone.js",
+  "author": {
+    "name": "Rich Harris"
+  },
+  "licenses": [
+    {
+      "type": "MIT"
+    }
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ractivejs/ractive-adaptors-backbone.git"
+  },
+  "peerDependencies": {
+    "backbone": ">= 1.0.0",
+    "ractive": ">= 0.3.9"
+  },
+  "scripts": {
+    "build": "rm -rf dist; gobble build dist",
+    "test": "jshint dist/ractive-adaptors-backbone.js && mocha",
+    "coverage": "(cp rdist/active-adaptors-backbone.js dist/ractive-adaptors-backbone.js~; jscoverage rdist/active-adaptors-backbone.js; mv dist/ractive-adaptors-backbone-cov.js dist/ractive-adaptors-backbone.js; mocha -R html-cov > coverage.html); mv dist/ractive-adaptors-backbone.js~ dist/ractive-adaptors-backbone.js",
+    "prepublish": "npm run build; uglifyjs -m < dist/ractive-adaptors-backbone.js > dist/ractive-adaptors-backbone.min.js"
+  },
+  "devDependencies": {
+    "backbone": ">= 1.0.0",
+    "chai": "^1.9.2",
+    "gobble": "^0.7.8",
+    "gobble-babel": "^4.0.1",
+    "gobble-esperanto-bundle": "^0.1.7",
+    "jshint": "^2.5.6",
+    "mocha": "^1.21.4",
+    "mocha-clean": "^0.2.1",
+    "mocha-repeat": "^0.1.0",
+    "proxyquire": "^1.0.1",
+    "ractive": ">= 0.3.9",
+    "uglify-js": "^2.4.15",
+    "underscore": "^1.7.0"
+  }
 }
